DM switching is now functional
Dependency on the "yourDeviceName" textblock removed. The App.xaml resource name is now used. Whenever a new button is selected, colour changes for the previous button and the current button. Messages now also clear.
The localhost button is now added automatically, and has to do so to declare currentButton to a value.