App.xaml username change implemented
First, I added a popup that is used to input the newUsername. Made DirectMessage.contactName settable. Now you can right click the username textblock to bring up a context menu that allows you to change your username, via the added dialog. changeUsername is a method you can use to change the username without a necessary dialog. It changes the networkManager entry for the selfContact, the resource entry, and yourDeviceName string binding. Lastly, I thought it'd be a great idea to add a "skype style" closing fakeout where the x button doesn't actually close the window.